backend/pyproject.toml
[tool.poetry]
name = "python-test-fastapi"
version = "0.0.5"
description = "Тестовый проект для fastAPI"
authors = ["kovalew.vladimir <kovalew.vladimir@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.103.2"
uvicorn = { extras = ["standart"], version = "^0.23.2" }
pydantic-settings = "^2.0.3"
python-jose = "^3.3.0"

python-ldap = "^3.4.3"
gssapi = "^1.8.3"

psycopg-binary = "^3.1.12"
psycopg = {extras = ["pool"], version = "^3.1.12"}
sqlalchemy = "^2.0.21"
alembic = "^1.12.0"

[tool.poetry.group.dev.dependencies]
black = "^23.9.1"
mypy = "^1.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pylint]
max-line-length = 120
disable = ["C0114", "C0115", "C0116"]
init-hook = 'import sys; sys.path.append("./src"); sys.path.append(".")'

[tool.mypy]
python_version = 3.12
# disallow_untyped_defs = "True"
